// app/utils/types/permission.ts
export enum Permission {
  Dashboard_View = 'view:dashboard',
  Orders_View = 'view:order',
  Production_View = 'view:production',
  Inventory_View = 'view:inventory',
  QualityControl_View = 'view:quantitycontrol',
  Logistics_View = 'view:logistics',
  Reports_View = 'view:reports',
  System_View = 'view:system'
}